ConnMan: lazy-unmount remote /flash and /storage before exiting
When net-booting LibreELEC (i.e., when /storage is mounted via NFS, NBD
or iSCSI), halting the system causes ConnMan to bring down the network
while processes (including ConnMan itself) still expect to be able to
write to /storage, resulting in those processes blocking and therefore
hanging the system indefinitely. Since nothing ought to be writing to
/storage by the time the network is being brought down, perform a lazy
unmount of /storage before stopping ConnMan.

Even though /flash is mounted read-only, do the same thing for /flash
too if it is mounted remotely: there's no reason not to.