init: only touch /dev/.kernel_ipconfig when network is configured manually
Kernel-level IP autoconfiguration (via the "ip" command line parameter)
allows network interfaces to be configured manually or via DHCP. The
kernel doesn't contain a DHCP client daemon, so if an interface is
configured via DHCP, it still needs to be managed in userspace after the
system finishes booting. Avoid touching /dev/.kernel_ipconfig in this
case when LibreELEC boots, to indicate to systemd that it should still
start a daemon that manages the network interface.